Works around bug in some files
ES has submitted files in which the date comes to this line not as a
string, but as an integer. This was causing an error. Since in all other
files this is a string, we're casting this to a string here.

TL;DR: we were getting int(201011) instead of "201011", which we usually
get.