Creates function to cleanup the METAR entries we get from REDEMET
The METAR we receive is not in the same format python-metar expects. The
timestamp and the '=' in the end of the line are not parsed by this lib.

This commit also skips some tests since the old parsing code is broken.
This tests will be replaced by the tests for the new parsing code.