Avoids second exception when there is one during capture
If there was an error in the capture process, the code then tried to
check if `data` was a list of non-zero size. The problem is that this
variable was not defined (since an exception occured when assigning to
it), which resulted in a second exception.
Now `data` is, by default, an empty list, which will not be saved to the
database. If the capture process works, then `data` will contain what
needs to be saved to the database and the code will work as expected.