Do not parse METAR entries with missing information
It turns out not only pressure information could be missing, but also
other fields. These fields are replaced by '////', so we do not try to
parse any line containing this string.

This has the disadvantage of skewing the averages a bit, but there seems
to be no way we can get the missing information through REDEMET.