Stop saving empty data to database
We should not save data for days with no weather information so the
crawler can find the date when it's trying to backfill empty data.
There was a bug that was creating almost empty entries in the database,
containing the date and the station code, but no weather information.
This commit should fix this issue and only write to the database when
there is weather data to be saved.